raspberrypi/requirements.txt
Adafruit-Blinka==8.18.0
adafruit-circuitpython-busdevice==5.2.4
adafruit-circuitpython-dht==3.7.8
adafruit-circuitpython-requests==1.13.1
adafruit-circuitpython-typing==1.9.1
Adafruit-DHT==1.4.0
Adafruit-PlatformDetect==3.45.1
Adafruit-PureIO==1.1.10
board==1.0
click==8.1.3
Flask==2.2.5
gpio==1.0.0
importlib-metadata==6.6.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
pyftdi==0.54.0
pyserial==3.5
pyusb==1.2.1
pywifi==1.1.12
RPi.GPIO==0.7.1
smbus==1.1.post2
spidev==3.6
typing-extensions==4.5.0
Werkzeug==2.2.3
zipp==3.15.0
